Fix breadcrumbs not showing when opening a timeline from case details (#115728)
